Assorted fixes to get the package to compile.

+++ src/static/include/scran/static/bindings.hpp
@@ -10,7 +10,7 @@ namespace sttc {
 
 typedef std::shared_ptr<tatami::NumericMatrix> Matrix;
 
-void PerCellQCMetrics(const Matrix& mat, std::vector<const double*> subsets, double* sums, int* detected, std::vector<double*> subset_proportions);
+void PerCellQCMetrics(const Matrix& mat, std::vector<const int*> subsets, double* sums, int* detected, std::vector<double*> subset_proportions);
 
 }
